Split the object from the static class
